+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/DriveConstants.java
@@ -88,9 +88,9 @@ public class DriveConstants {
 
      */
     public static double MAX_VEL = 52.48291908330528;
-    public static double MAX_ACCEL = 52.48291908330528;
-    public static double MAX_ANG_VEL = Math.toRadians(167.05832);
-    public static double MAX_ANG_ACCEL = Math.toRadians(167.05832);
+    public static double MAX_ACCEL = 300;
+    public static double MAX_ANG_VEL = Math.toRadians(300);
+    public static double MAX_ANG_ACCEL = Math.toRadians(300);
 
 
     public static double encoderTicksToInches(double ticks) {
