+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/opmode/EXP_RR_Auto_Red_Right.java
@@ -14,39 +14,40 @@ public class EXP_RR_Auto_Red_Right extends LinearOpMode {
     public void runOpMode() {
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
 
+        waitForStart();
         TrajectorySequence myTrajectory = drive.trajectorySequenceBuilder(new Pose2d())
-                .forward(57)
-                .turn(Math.toRadians(40))
+                .forward(74)
+                .turn(Math.toRadians(60))
                 //Place Cone
-                .turn(Math.toRadians(-121))
+                .turn(Math.toRadians(-138))
                 .forward(20)
                 //Grab Cone
                 .back(20)
-                .turn(Math.toRadians(121))
+                .turn(Math.toRadians(-55))
                 //place cone
-                .turn(Math.toRadians(-121))
+                .turn(Math.toRadians(55))
                 .forward(20)
                 //Grab Cone
                 .back(20)
-                .turn(Math.toRadians(121))
+                .turn(Math.toRadians(-55))
                 //place cone
-                .turn(Math.toRadians(-121))
+                .turn(Math.toRadians(55))
                 .forward(20)
                 //Grab Cone
                 .back(20)
-                .turn(Math.toRadians(121))
+                .turn(Math.toRadians(-55))
                 //place cone
-                .turn(Math.toRadians(-121))
+                .turn(Math.toRadians(55))
                 .forward(20)
                 //Grab Cone
                 .back(20)
-                .turn(Math.toRadians(121))
+                .turn(Math.toRadians(-55))
                 //place cone
-                .turn(Math.toRadians(-121))
+                .turn(Math.toRadians(55))
                 .forward(20)
                 //Grab Cone
                 .back(20)
-                .turn(Math.toRadians(121))
+                .turn(Math.toRadians(-55))
                 .build();
 
         waitForStart();
